Compiler warnings consolidation.

+++ src/lib/parser/node/BlockNode.cpp
@@ -119,7 +119,7 @@ void BlockNode::execute(IndentWriter &writer, PugModel &model, PugTemplate &tmpl
         writer.prettyIndent(1, true);
     }
 
-    for (int i = 0; i < nodes_.size(); ++i)
+    for (size_t i = 0; i < nodes_.size(); ++i)
     {
         // Pretty print text
         if (writer.isPp() && i > 0 && !writer.isEscape() && nodes_[i]->getType() == e_TextNode &&
